+++ 新浪微博/Classes/Other/Lib/MJPhotoBrowser/MJPhotoToolbar.m
@@ -80,7 +80,7 @@
     _currentPhotoIndex = currentPhotoIndex;
     
     // 更新页码
-    _indexLabel.text = [NSString stringWithFormat:@"%d / %d", _currentPhotoIndex + 1, _photos.count];
+    _indexLabel.text = [NSString stringWithFormat:@"%d / %d", (int)_currentPhotoIndex + 1, (int)_photos.count];
     if (_photos.count) {
         
         MJPhoto *photo = _photos[_currentPhotoIndex];
